Address workflow validation review comments
Make release asset uploads idempotent by passing `--clobber` to
`gh release upload` in the generated release workflow.

Expand the test package documentation and move generated-file parsing and
workflow contract assertions into dedicated helper modules so
`tests/test_template.py` remains focused on top-level test cases.
