+++ template/.github/workflows/release.yml.jinja
@@ -62,6 +62,6 @@ jobs:
         run: |
           set -eu
           find dist/wheels-* -type f -name "*.whl" -print0 | \
-            xargs -0 -r gh release upload "${{ "{{" }} github.ref_name {{ "}}" }}"
+            xargs -0 -r gh release upload "${{ "{{" }} github.ref_name {{ "}}" }}" --clobber
         env:
           GITHUB_TOKEN: ${{ "{{" }} secrets.GITHUB_TOKEN {{ "}}" }}
