+++ tests/helpers/__init__.py
@@ -1,0 +1,1 @@
+"""Shared helper modules for rendered Copier template tests."""
